requirements.txt
Django==1.9.4
django-bootstrap-pagination==1.6.1
django-bootstrap3==7.0.0
djangorestframework==3.3.3
twitter==1.17.1